requirements.txt
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Django==3.1.7
django-storages==1.11.1
google-api-core==1.26.0
google-auth==1.27.0
google-cloud-bigquery==2.10.0
google-cloud-core==1.6.0
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.0
idna==2.10
packaging==20.9
proto-plus==1.14.2
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
